Disable SSL validation.
The Instagram API is signed with a GeoTrust Global CA. Unfortunately
this CA root certificate is not part of the default cacert.txt file
that ships with httplib2.

Rather than ship the certificate file, let's just ignore the SSL
validation issue.

Once the upstream bug is fixed,
 http://code.google.com/p/httplib2/issues/detail?id=156
we can revert this change.